requirements.txt
asgiref==3.7.2
bleach==6.1.0
Django==5.0
django-markdownfield==0.11.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Markdown==3.5.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
webencodings==0.5.1
